Add Helper class to hold helper methods that simplify and abstract away details, but also have performance advantages
